Added new GitHubStatus ADDED_RENAMED to fix false negative detection
There was a false negative detection if files were added and renamed in
the same milestone. The issue is that the RENAMED files have no link to
the old classes. The solution is to parse the old class name from the
diff patch of the file and try to find a match in the complete file set.
If a match is found we change the state to ADDED_RENAMED and apply the
minimal code coverage for added files instead of modified files.